Move evaluator/mod.rs code to (new file) functions.rs
Last of the migration of code out of mod.rs files.
